Bug fix in tagCells.
We were skipping the undivided difference calculation in directions that would not be immediately refined (directions whose refRatio=1). Now, we only skip if those directions will never be refined.

For example, suppose refRatio on lev 0 is (2,1), refRatio on lev 1 is (2, 1), and refRatio on lev 2 is (4,4). Also, suppose on lev 0, we are tagging cells. We would have skipped checking undivided differences in the y-direction because it would not have been immediately refined. With this fix, we will check the y-dir differences because refining levs 0 and 1 will help resolve the flow.